Move to the class
